Refactor: Move persistent resources to a new API namespace
OK, fine. Accessing the fog parameters will be rather awkward if they're kept in the top-level client module, so let's take one step towards streamlining the resource access without going overboard.
